Implementation of ceil of integer logarithm
Implement integer logarithm based on the concept of binary search, in
order to return the ceiling of integer logarithm, we add 1 at the last
instruction.

By the way, in order to handle to case when x is the power of 2, we need
to substract x by 1 before doing binary search, but it'll cause an
underflow in the case of 0 which leads to 0xFFFFFFFF after decrease by
1, so we use the instruction "x -= (!!x)" to prevent x from minus 1 in
the case of x=0.