Modify xt_update function to fix unbalance case
In the origin implementation of xtree, it's likely to generate an
unbalanced xtree in which the difference between the left subtree
height and the right subtree height become larger than 2 without
correctly modify itself. The main reason is the original behavior when
trying to balance the subtree, we might remove the longest subtree from
left to right or from right to left, so the update operation will be in
vain.

We treat it as the case in AVL tree so we can seperate the unbalance
cases into RR, RL, LR and LL. If the unbalance case is RL or LR, we have
to rotate twice just as what we did in AVL tree, we can utilize the
"xt_balance" factor to determine whether it's RL or LR.